requirements.txt
# --------- pytorch --------- #
torch>=1.8.1
torchvision>=0.9.1
pytorch-lightning>=1.3.8

# --------- hydra --------- #
hydra-core==1.1.0
hydra-colorlog==1.1.0
hydra-optuna-sweeper==1.1.0
# hydra-ax-sweeper
# hydra-ray-launcher
# hydra-submitit-launcher

# --------- loggers --------- #
wandb~=0.12.2
# neptune-client
# mlflow
# comet-ml
# torch_tb_profiler

# --------- linters --------- #
pre-commit      # hooks for applying linters on commit
black           # code formatting
isort           # import sorting
flake8          # code analysis

# --------- others --------- #
python-dotenv~=0.19.0
rich~=10.9.0
pytest~=6.2.5
sh~=1.14.2
scikit-learn~=0.24.2
seaborn~=0.11.2
jupyterlab      # better jupyter notebooks
pudb            # debugger

spacy~=3.1.3
omegaconf~=2.1.1
torchmetrics~=0.5.1
matplotlib~=3.4.3
numpy~=1.19.5
packaging~=21.0
tqdm~=4.62.2
datasets
transformers
allennlp-models